refactor codelenses
